requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai
openai
faiss-cpu
youtube-transcript-api
pytube
tiktoken
textwrap
